+++ include/util/util.hpp
@@ -15,9 +15,9 @@ struct MainContext {
     MainContext();
 
     bool is_verbose;
-    int32_t debug_code;
-    int32_t optim_1_mask;
-    int32_t optim_2_code;
+    uint8_t debug_code;
+    uint8_t optim_1_mask;
+    uint8_t optim_2_code;
     std::string filename;
     std::vector<std::string> includedirs;
     std::vector<std::string> args;
