+++ src/optimization/reg_alloc.cpp
@@ -12,7 +12,7 @@ static std::unique_ptr<RegAllocContext> context;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-void register_allocation(AsmProgram* /*node*/, int32_t /*optim_2_code*/) {
+void register_allocation(AsmProgram* /*node*/, uint8_t /*optim_2_code*/) {
     context = std::make_unique<RegAllocContext>();
     context.reset();
 }
